+++ cmd/cert-controller-manager-next-generation/.import-restrictions
@@ -1,7 +1,7 @@
 rules:
   - selectorRegexp: github[.]com/gardener/cert-management
     allowedPrefixes:
-      - github.com/gardener/cert-management/cmd/certman2
+      - github.com/gardener/cert-management/cmd/cert-controller-manager-next-generation
       - github.com/gardener/cert-management/pkg/apis
       - github.com/gardener/cert-management/pkg/certman2
       - github.com/gardener/cert-management/pkg/shared
